+++ Assets/Script/Player/PlayerHealth.cs
@@ -6,14 +6,14 @@ public class PlayerHealth : CreatureHealth
     [SerializeField] private HealthUI healthUI;
     
     public bool isDeflecting = false;
-    public override void GetDamage(int damage)
+    public override void GetDamage(float damage)
     {
         if(isDeflecting) return;
         base.GetDamage(damage);
         healthUI.ChangeHeartsCount(damage * -1);
     }
 
-    public override void GetHeal(int heal)
+    public override void GetHeal(float heal)
     {
         base.GetHeal(heal);
         healthUI.ChangeHeartsCount(heal);
